rename arrops public header to distopia.h
