+++ src/lib/src/calc_angles.cpp
@@ -7,11 +7,10 @@
 #include <immintrin.h>
 #include <iostream>
 
-#include "arrops.h"
+#include "distopia.h"
 #include "compiler_hints.h"
 #include "distopia_type_traits.h"
 #include "vector_triple.h"
-
 #include "ortho_box.h"
 
 namespace {
